moving solana mcms out of internal
