+++ deployment/common/changeset/solana/transfer_ownership_test.go
@@ -15,8 +15,8 @@ import (
 
 	"github.com/smartcontractkit/chainlink/deployment"
 	commonchangeset "github.com/smartcontractkit/chainlink/deployment/common/changeset"
-	internalsolana "github.com/smartcontractkit/chainlink/deployment/common/changeset/internal/solana"
 	solanachangesets "github.com/smartcontractkit/chainlink/deployment/common/changeset/solana"
+	solanaMCMS "github.com/smartcontractkit/chainlink/deployment/common/changeset/solana/mcms"
 	"github.com/smartcontractkit/chainlink/deployment/common/changeset/state"
 	"github.com/smartcontractkit/chainlink/deployment/common/proposalutils"
 	commontypes "github.com/smartcontractkit/chainlink/deployment/common/types"
@@ -68,7 +68,7 @@ func deployMCMS(t *testing.T, env deployment.Environment, selector uint64) *stat
 		TimelockMinDelay: big.NewInt(1),
 	}
 
-	chainState, err := internalsolana.DeployMCMSWithTimelockProgramsSolana(env, solanaChain, addressBook, mcmsConfig)
+	chainState, err := solanaMCMS.DeployMCMSWithTimelockProgramsSolana(env, solanaChain, addressBook, mcmsConfig)
 	require.NoError(t, err)
 	err = env.ExistingAddresses.Merge(addressBook)
 	require.NoError(t, err)
